awf: fix(#698): don't misclassify a dead agent-service idle-timeout as a
Authored by AWF workspace ws_61dfc66e1e8b40f4aaddf398 (agent: codex).
